# frontend/docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AI Chat API",
  "description": "Backend API for the AI companion/chat application: authentication, AI characters, conversations, messages, memories, and streamed AI replies (SSE). All endpoints except `/auth/register` and `/auth/login` require a Bearer token.",
  "version": "0.1.0"
 },
 "paths": {
  "/auth/register": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Register a new account",
    "description": "Creates a user account. The password is hashed with Argon2id and never stored in plaintext. Usernames are unique.",
    "operationId": "register_auth_register_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RegisterRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/auth/login": {
   "post": {
    "tags": [
     "auth"
    ],
    "summary": "Log in and obtain an access token",
    "description": "Exchanges username/password for a signed JWT access token. Send it as `Authorization: Bearer <token>` on protected endpoints.",
    "operationId": "login_auth_login_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/LoginResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/characters": {
   "get": {
    "tags": [
     "characters"
    ],
    "summary": "List active AI characters",
    "operationId": "list_characters_characters_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/CharacterRead"
         },
         "type": "array",
         "title": "Response List Characters Characters Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/characters/{character_id}": {
   "get": {
    "tags": [
     "characters"
    ],
    "summary": "Get one active AI character",
    "operationId": "get_character_characters__character_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "character_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Character Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CharacterRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/conversations": {
   "get": {
    "tags": [
     "conversations"
    ],
    "summary": "List the user's conversations",
    "operationId": "list_conversations_conversations_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ConversationRead"
         },
         "title": "Response List Conversations Conversations Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "conversations"
    ],
    "summary": "Create a new conversation",
    "description": "Starts a conversation with an AI character. The model defaults to the user's default model, then to the first active model.",
    "operationId": "create_conversation_conversations_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConversationCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationRead"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/conversations/{conversation_id}/messages": {
   "get": {
    "tags": [
     "messages"
    ],
    "summary": "List a conversation's messages",
    "description": "Returns messages in chronological order. Use `before_id` (exclusive) with `limit` to page backwards through history.",
    "operationId": "list_messages_conversations__conversation_id__messages_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Conversation Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "before_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer",
         "minimum": 1
        },
        {
         "type": "null"
        }
       ],
       "title": "Before Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/MessageRead"
         },
         "title": "Response List Messages Conversations  Conversation Id  Messages Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "messages"
    ],
    "summary": "Send a message and stream the AI reply (Server-Sent Events)",
    "description": "Persists the user message, then streams the AI reply incrementally as SSE. Events: `message.created` (the persisted user message), repeated `message.delta` (incremental text), `message.completed` (the persisted assistant message + token usage), or `error` if the AI provider fails.",
    "operationId": "send_message_conversations__conversation_id__messages_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "conversation_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "title": "Conversation Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/MessageCreate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "SSE stream",
      "content": {
       "text/event-stream": {}
      }
     },
     "401": {
      "description": "Missing or invalid access token"
     },
     "404": {
      "description": "Conversation not found or not owned by the caller"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "CharacterBrief": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "CharacterBrief"
   },
   "CharacterRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "description",
     "avatar_url",
     "created_at"
    ],
    "title": "CharacterRead"
   },
   "ConversationCreate": {
    "properties": {
     "character_id": {
      "type": "integer",
      "minimum": 1.0,
      "title": "Character Id"
     },
     "model_id": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Model Id",
      "description": "Defaults to the user's default model"
     },
     "title": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "type": "object",
    "required": [
     "character_id"
    ],
    "title": "ConversationCreate",
    "description": "Body for POST /conversations."
   },
   "ConversationRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "character": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CharacterBrief"
       },
       {
        "type": "null"
       }
      ]
     },
     "model": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/ModelBrief"
       },
       {
        "type": "null"
       }
      ]
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "last_message_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Message At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "character",
     "model",
     "title",
     "status",
     "created_at",
     "updated_at",
     "last_message_at"
    ],
    "title": "ConversationRead"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LoginRequest": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 32,
      "minLength": 1,
      "title": "Username"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 1,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "LoginRequest",
    "description": "Body for POST /auth/login."
   },
   "LoginResponse": {
    "properties": {
     "access_token": {
      "type": "string",
      "title": "Access Token"
     },
     "token_type": {
      "type": "string",
      "const": "bearer",
      "title": "Token Type",
      "default": "bearer"
     },
     "expires_in": {
      "type": "integer",
      "title": "Expires In",
      "description": "Token lifetime in minutes"
     },
     "user": {
      "$ref": "#/components/schemas/UserRead"
     }
    },
    "type": "object",
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "LoginResponse",
    "description": "Successful login: a bearer access token plus the user profile."
   },
...
   "MessageCreate": {
    "properties": {
     "content": {
      "type": "string",
      "maxLength": 32000,
      "minLength": 1,
      "title": "Content"
     },
     "reply_to_id": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1,
        "title": "Reply To Id"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reply To Id"
     }
    },
    "type": "object",
    "required": [
     "content"
    ],
    "title": "MessageCreate",
    "description": "Body for POST /conversations/{id}/messages (starts a streamed AI reply)."
   },
   "MessageRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "conversation_id": {
      "type": "integer",
      "title": "Conversation Id"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "content": {
      "type": "string",
      "title": "Content"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "reply_to_id": {
      "anyOf": [
       {
        "type": "integer",
        "title": "Reply To Id"
       },
       {
        "type": "null"
       }
      ],
      "title": "Reply To Id"
     }
    },
    "type": "object",
    "required": [
     "id",
     "conversation_id",
     "role",
     "content",
     "created_at",
     "reply_to_id"
    ],
    "title": "MessageRead"
   },
   "ModelBrief": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "model_name": {
      "type": "string",
      "title": "Model Name"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "id",
     "model_name",
     "display_name"
    ],
    "title": "ModelBrief"
   },
...
   "RegisterRequest": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 32,
      "minLength": 3,
      "pattern": "^[A-Za-z0-9_.-]+$",
      "title": "Username"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "minLength": 8,
      "title": "Password"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     }
    },
    "type": "object",
    "required": [
     "username",
     "password"
    ],
    "title": "RegisterRequest",
    "description": "Body for POST /auth/register."
   },
   "UserRead": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "display_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Display Name"
     },
     "role": {
      "type": "string",
      "title": "Role"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "default_model_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Model Id"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "last_login_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Login At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "username",
     "display_name",
     "role",
     "status",
     "default_model_id",
     "created_at",
     "last_login_at"
    ],
    "title": "UserRead",
    "description": "Public user profile. Never contains the password hash."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "JWT access token from POST /auth/login",
    "scheme": "bearer"
   }
  }
 }
}